fix(main_messages) Code Refactor (#213)
Refactor for code cleanup.
DocString improvements.
Most Pylint issues resolved.

d_rats/ui/messages_tab.py:
  Remove MessageFolderInfo, MessageFolders, and MessageList classes.

d_rats/ui/message_folder_info.py:
d_rats/ui/messages_tab.py:
d_rats/ui/message_list.py:
  New modules for the above classes.
