+++ d_rats/ui/main_messages.py
@@ -1,7 +1,5 @@
 #!/usr/bin/python
 '''Main Messages'''
-# pylint wants a max of 1000 lines per module
-# pylint: disable=too-many-lines
 #
 # Copyright 2009 Dan Smith <dsmith@danplanet.com>
 # Copyright 2021-2022 John. E. Malmberg - Python3 Conversion
@@ -29,23 +27,18 @@ import shutil
 import random
 
 from glob import glob
-from datetime import datetime
-from configparser import ConfigParser
-from configparser import DuplicateSectionError
-from configparser import NoOptionError
-from configparser import NoSectionError
 
 import gi
 gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk
-from gi.repository import Gdk
-from gi.repository import GLib
-from gi.repository import GObject
-from gi.repository import Pango
 
-from d_rats.ui.main_common import MainWindowElement, MainWindowTab
-from d_rats.ui.main_common import prompt_for_station, \
-    display_error, prompt_for_string, set_toolbar_buttons
+from d_rats.ui.main_common import MainWindowTab
+from d_rats.ui.main_common import prompt_for_station
+from d_rats.ui.main_common import display_error
+from d_rats.ui.main_common import set_toolbar_buttons
+from d_rats.ui.message_folder_info import MessageFolderInfo
+from d_rats.ui.message_folders import MessageFolders
+from d_rats.ui.message_list import MessageList
 
 from d_rats import inputdialog
 from d_rats import formgui
@@ -53,21 +46,9 @@ from d_rats import signals
 from d_rats import msgrouting
 
 
-_FOLDER_CACHE = {}
-
 if not '_' in locals():
     import gettext
     _ = gettext.gettext
-
-BASE_FOLDERS = [_("Inbox"), _("Outbox"), _("Sent"), _("Trash"), _("Drafts")]
-
-
-class MainMessageException(Exception):
-    '''Generic Main Message Exception.'''
-
-
-class FolderError(MainMessageException):
-    '''Error accessing a folder.'''
 
 
 def mkmsgid(callsign):
@@ -81,1103 +62,6 @@ def mkmsgid(callsign):
     '''
     r_num = random.SystemRandom().randint(0, 100000)
     return "%s.%x.%x" % (callsign, int(time.time()) - 1114880400, r_num)
-
-
-class MessageFolderInfo():
-    '''
-    Message Folder Info.
-
-    :param folder_path: Folder to operate on
-    :type folder_path: str
-    '''
-
-    def __init__(self, folder_path):
-        self._path = folder_path
-
-        self.logger = logging.getLogger("MessageFolderInfo")
-        if folder_path in _FOLDER_CACHE:
-            self._config = _FOLDER_CACHE[folder_path]
-        else:
-            self._config = ConfigParser()
-            reg_path = os.path.join(self._path, ".db")
-            if os.path.exists(reg_path):
-                self._config.read(reg_path)
-            self._save()
-            _FOLDER_CACHE[folder_path] = self._config
-
-    def _save(self):
-        reg_path = os.path.join(self._path, ".db")
-        file_handle = open(reg_path, "w")
-        self._config.write(file_handle)
-        file_handle.close()
-
-    def name(self):
-        '''
-        Folder Name.
-
-        :returns: Current Folder name
-        :rtype: str
-        '''
-        return os.path.basename(self._path)
-
-    def _set_prop(self, filename, prop, value):
-        filename = os.path.basename(filename)
-
-        if not self._config.has_section(filename):
-            self._config.add_section(filename)
-
-        self._config.set(filename, prop, value)
-        self._save()
-
-    def _get_prop(self, filename, prop):
-        filename = os.path.basename(filename)
-
-        try:
-            return self._config.get(filename, prop)
-        except (NoOptionError, NoSectionError):
-            return _("Unknown")
-
-    def get_msg_subject(self, filename):
-        '''
-        Get message subject.
-
-        :param filename: Filename for message
-        :type filename: str
-        :returns: Subject of message
-        :rtype: str
-        '''
-        return self._get_prop(filename, "subject")
-
-    def set_msg_subject(self, filename, subject):
-        '''
-        Set message subject.
-
-        :param filename: Filename for message
-        :type filename: str
-        :param subject: Subject for message
-        :type subject: str
-        '''
-        self._set_prop(filename, "subject", subject)
-
-    def get_msg_type(self, filename):
-        '''
-        Get message type.
-
-        :param filename: Filename for message
-        :type filename: str
-        :returns: Message Type
-        :rtype: str
-        '''
-        return self._get_prop(filename, "type")
-
-    def set_msg_type(self, filename, msg_type):
-        '''
-        Set message type.
-
-        :param filename: Filename for message
-        :type filename: str
-        :param msg_type: Type of message
-        :type msg_type: str
-        '''
-        self._set_prop(filename, "type", msg_type)
-
-    def get_msg_read(self, filename):
-        '''
-        Get message read status.
-
-        :param filename: Filename for message
-        :type filename: str
-        :returns: True if message has been read
-        :rtype: bool
-        '''
-        val = self._get_prop(filename, "read")
-        return val == "True"
-
-    def set_msg_read(self, filename, read):
-        '''
-        Set message read status.
-
-        :param filename: Filename of message
-        :type filename: str
-        :param read: True for message to be marked read
-        :type read: bool
-        '''
-        self._set_prop(filename, "read", str(read))
-
-    def get_msg_sender(self, filename):
-        '''
-        Get the message sender.
-
-        :param filename: Filename of message
-        :type filename: str
-        :returns: Sender of message
-        :rtype: str
-        '''
-        return self._get_prop(filename, "sender")
-
-    def set_msg_sender(self, filename, sender):
-        '''
-        Set message sender.
-
-        :param filename: Filename for message
-        :type filename: str
-        :param sender: Sender of message
-        :type sender: str
-        '''
-        self._set_prop(filename, "sender", sender)
-
-    def get_msg_recip(self, filename):
-        '''
-        Get the message recipient
-
-        :param filename: Filename for message
-        :type filename: str
-        :returns: Message recipient
-        :rtype: str
-        '''
-        return self._get_prop(filename, "recip")
-
-    def set_msg_recip(self, filename, recip):
-        '''
-        Set the message recipient.
-
-        :param filename: Filename for message
-        :type filename: str
-        :param recip: Recipient
-        :type recip: str
-        '''
-        self._set_prop(filename, "recip", recip)
-
-    def subfolders(self):
-        '''
-        Get the subfolders.
-
-        :returns: subfolders of folder
-        :rtype: list[:class:`MessageFolderInfo`]
-        '''
-        info = []
-
-        entries = glob(os.path.join(self._path, "*"))
-        for entry in sorted(entries):
-            if entry in (".", ".."):
-                continue
-            if os.path.isdir(entry):
-                info.append(MessageFolderInfo(entry))
-
-        return info
-
-    def files(self):
-        '''
-        List files.
-
-        :returns: files in the folder.
-        :rtype: list[str]
-        '''
-        file_list = glob(os.path.join(self._path, "*"))
-        return [x_file for x_file in file_list
-                if os.path.isfile(x_file) and not x_file.startswith(".")]
-
-    def get_subfolder(self, name):
-        '''
-        Get subfolder information.
-
-        :param name: Subfolder name
-        :type name: str
-        :returns: Subfolder information
-        :rtype: :class:`MessageFolderInfo`
-        '''
-        for folder in self.subfolders():
-            if folder.name() == name:
-                return folder
-
-        return None
-
-    def create_subfolder(self, name):
-        '''
-        Create a subfolder by name
-
-        :param name: Subfolder name
-        :type name: str
-        :returns: subfolder information
-        :rtype: :class:`MessageFolderInfo`
-        '''
-        path = os.path.join(self._path, name)
-        try:
-            os.mkdir(path)
-        except OSError as err:
-            if err.errno != 17:  # File or directory exists
-                raise
-        return MessageFolderInfo(path)
-
-    def delete_self(self):
-        '''Delete Self.'''
-        try:
-            os.remove(os.path.join(self._path, ".db"))
-        except OSError:
-            pass # Don't freak if no .db
-        os.rmdir(self._path)
-
-    def create_msg(self, name):
-        '''
-        Create a message.
-
-        Store the message name in the users configuration data
-
-        :param name: Name for message path
-        :type name: str
-        :returns: Path for message
-        :rtype: str
-        :raises: DuplicateSectionError if the section already exists
-        '''
-        exists = os.path.exists(os.path.join(self._path, name))
-        try:
-            self._config.add_section(name)
-        except DuplicateSectionError as err:
-            if exists:
-                raise err
-
-        return os.path.join(self._path, name)
-
-    def delete(self, filename):
-        '''
-        Delete a file.
-
-        :param filename: filename to delete
-        :type filename: str
-        '''
-        filename = os.path.basename(filename)
-        self._config.remove_section(filename)
-        os.remove(os.path.join(self._path, filename))
-
-    def rename(self, new_name):
-        '''
-        Rename path
-
-        :param new_name: New name for path
-        :type new_name: str
-        '''
-        new_path = os.path.join(os.path.dirname(self._path), new_name)
-        self.logger.info("Renaming %s -> %s", self._path, new_path)
-        os.rename(self._path, new_path)
-        self._path = new_path
-
-    def __str__(self):
-        return self.name()
-
-
-# pylint wants at least 2 public methods per class.
-# pylint: disable=too-few-public-methods
-class MessageInfo():
-    '''Message information.
-
-    :param filename: Filename of message
-    :type filename: str
-    :param info: Information about the message
-    :type info: :class:`MessageFolderInfo`
-    '''
-
-    def __init__(self, filename, info):
-        self.filename = filename
-        self.info = info
-
-
-class MessageFolders(MainWindowElement):
-    '''
-    Message Folders.
-
-    :param wtree: Window tree
-    :type wtree: :class:`Gtk.GtkNotebook`
-    :param config: Configuration data
-    :type config: :class:`DratsConfig`
-    '''
-
-    __gsignals__ = {
-        "user-selected-folder" : (GObject.SignalFlags.RUN_LAST,
-                                  GObject.TYPE_NONE,
-                                  (GObject.TYPE_STRING,))
-        }
-
-    # MessageFolders
-    def __init__(self, wtree, config):
-        MainWindowElement.__init__(self, wtree, config, "msg")
-
-        self.logger = logging.getLogger("MessageFolders")
-        folderlist = self._get_widget("folderlist")
-
-        store = Gtk.TreeStore(GObject.TYPE_STRING, GObject.TYPE_OBJECT)
-        folderlist.set_model(store)
-        folderlist.set_headers_visible(False)
-        folderlist.enable_model_drag_dest([("text/d-rats_message", 0, 0)],
-                                          Gdk.DragAction.DEFAULT)
-        folderlist.connect("drag-data-received", self._dragged_to)
-        folderlist.connect("button_press_event", self._mouse_cb)
-        folderlist.connect_after("move-cursor", self._move_cursor)
-
-        col = Gtk.TreeViewColumn("", Gtk.CellRendererPixbuf(), pixbuf=1)
-        folderlist.append_column(col)
-
-        rnd = Gtk.CellRendererText()
-        #rnd.set_property("editable", True)
-        rnd.connect("edited", self._folder_rename, store)
-        col = Gtk.TreeViewColumn("", rnd, text=0)
-        folderlist.append_column(col)
-
-        self.folder_pixbuf = self._config.ship_img("folder.png")
-
-        self._ensure_default_folders()
-        for folder in self.get_folders():
-            self._add_folders(store, None, folder)
-
-    def _folders_path(self):
-        '''
-        Folders Path.
-
-        :returns: The folder path
-        :rtype: str
-        '''
-        path = os.path.join(self._config.platform.config_dir(), "messages")
-        if not os.path.isdir(path):
-            os.makedirs(path)
-        return path
-
-    @staticmethod
-    def _create_folder(root, name):
-        '''
-        Create folder.
-
-        :param root: Parent directory
-        :type root: :class:`MessageFolderInfo`
-        :param name: Folder name
-        :type name: str
-        :returns: Message folder info for child directory
-        :rtype: :class:`MessageFolderInfo`
-        :raises: :class:`FolderError` if folder can not be created
-        '''
-        # python3 can create directory and parent directories in one call.
-        info = root
-        for dir_element in name.split(os.sep)[:-1]:
-            info = info.get_subfolder(dir_element)
-            if not info:
-                break
-
-        try:
-            return info.create_subfolder(os.path.basename(name))
-        except OSError as err:
-            if err.errno != 17:  # File or directory exists
-                raise FolderError(
-                    "Intermediate folder of %s does not exist %s" %
-                    (name, err))
-
-    def create_folder(self, name):
-        '''
-        Create a folder.
-
-        Does not appear to be called anywhere.
-
-        :param name: Folder name
-        :type name: str
-        :raises: :class:`FolderError` if folder cannot be created
-        '''
-        root = MessageFolderInfo(self._folders_path())
-        self._create_folder(root, name)
-
-    def get_folders(self):
-        '''
-        Get Folders.
-
-        :returns: Message folders information
-        :rtype: list[:class:`MessageFolderInfo`]
-        '''
-        return MessageFolderInfo(self._folders_path()).subfolders()
-
-    def get_folder(self, name):
-        '''
-        Get Folder by name.
-
-        :param name: Folder name
-        :type name: str
-        :returns: Message folder information
-        :rtype: :class:`MessageFolderInfo`
-        '''
-        return MessageFolderInfo(os.path.join(self._folders_path(), name))
-
-    @staticmethod
-    def _get_folder_by_iter(store, msg_iter):
-        els = []
-        while msg_iter:
-            els.insert(0, store.get(msg_iter, 0)[0])
-            msg_iter = store.iter_parent(msg_iter)
-
-        return os.sep.join(els)
-
-    def select_folder(self, folder):
-        '''
-        Select a folder by path.
-
-        i.e. Inbox/Subfolder
-        NB: Subfolders currently not supported.
-
-        :param folder: Folder name to select
-        :type folder: str
-        '''
-        view = self._get_widget("folderlist")
-        store = view.get_model()
-
-        msg_iter = store.get_iter_first()
-        while msg_iter:
-            fqname = self._get_folder_by_iter(store, msg_iter)
-            if fqname == folder:
-                view.set_cursor(store.get_path(msg_iter))
-                self.emit("user-selected-folder", fqname)
-                break
-
-            msg_iter = store.iter_next(msg_iter)
-
-    def _ensure_default_folders(self):
-        root = MessageFolderInfo(self._folders_path())
-
-        for folder in BASE_FOLDERS:
-            try:
-                info = self._create_folder(root, folder)
-                self.logger.info("_ensure_default_folders %s",
-                                 info.subfolders())
-            except FolderError:
-                pass
-
-    def _add_folders(self, store, msg_iter, root):
-        msg_iter = store.append(msg_iter, (root.name(), self.folder_pixbuf))
-        for info in root.subfolders():
-            self._add_folders(store, msg_iter, info)
-
-    @staticmethod
-    def _get_selected_folder(view, event):
-        if event.window == view.get_bin_window():
-            x_coord, y_coord = event.get_coords()
-            pathinfo = view.get_path_at_pos(int(x_coord), int(y_coord))
-            if pathinfo is None:
-                return view.get_model(), None
-            view.set_cursor_on_cell(pathinfo[0], None, None, False)
-
-        return view.get_selection().get_selected()
-
-    def _menu_handler(self, action, store, msg_iter, _view):
-        '''
-        Menu activate handler.
-
-        :param action: Action that was signaled
-        :type action: :class:`Gtk.Action`
-        :param store: Tree store widget
-        :type store: :class:`Gtk.TreeStore`
-        :param msg_iter: Message iterator
-        :type msg_iter: :class:`Gtk.TreeIter`
-        :param view: Mouse button widget
-        :type view: :class:`Gtk.TreeView`
-        '''
-        action_name = action.get_name()
-
-        if action_name == "delete":
-            info = self.get_folder(self._get_folder_by_iter(store, msg_iter))
-            try:
-                info.delete_self()
-            except OSError as err:
-                display_error("Unable to delete folder: %s" % err)
-                return
-            store.remove(msg_iter)
-        elif action_name == "create":
-            store.insert(msg_iter, 0, ("New Folder", self.folder_pixbuf))
-            parent = self.get_folder(self._get_folder_by_iter(store, msg_iter))
-            self._create_folder(parent, "New Folder")
-        elif action_name == "rename":
-            info = self.get_folder(self._get_folder_by_iter(store, msg_iter))
-
-            new_text = prompt_for_string("Rename folder `%s' to:" % info.name(),
-                                         orig=info.name())
-            if not new_text:
-                return
-            if new_text == info.name():
-                return
-
-            try:
-                info.rename(new_text)
-            except OSError as err:
-                display_error("Unable to rename: %s" % err)
-                return
-
-            store.set(msg_iter, 0, new_text)
-
-    def _select_folder(self, view, event):
-        store, msg_iter = self._get_selected_folder(view, event)
-        if not msg_iter:
-            return
-        self.emit("user-selected-folder",
-                  self._get_folder_by_iter(store, msg_iter))
-
-    def _move_cursor(self, view, _step, _direction):
-        '''
-        Move cursor move-cursor handler.
-
-        :param view: TreeView widget
-        :type view: :class:`Gtk.TreeView`
-        :param _step: The granularity of the move
-        :type _step: :class:`Gtk.MovementStep`
-        :param _direction:  Direction to move
-        :type _direction: int
-        '''
-        (store, msg_iter) = view.get_selection().get_selected()
-        self.emit("user-selected-folder",
-                  self._get_folder_by_iter(store, msg_iter))
-
-    # pylint wants a max of 15 local variables
-    # pylint: disable=too-many-locals
-    def _folder_menu(self, view, event):
-        x_coord = int(event.x)
-        y_coord = int(event.y)
-        pthinfo = view.get_path_at_pos(x_coord, y_coord)
-        if pthinfo is not None:
-            path, col, _cellx, _celly = pthinfo
-            view.grab_focus()
-            view.set_cursor(path, col, 0)
-
-        xml = """
-<ui>
-  <popup name="menu">
-    <menuitem action="delete"/>
-    <menuitem action="create"/>
-    <menuitem action="rename"/>
-  </popup>
-</ui>
-"""
-        store, folder_iter = self._get_selected_folder(view, event)
-        folder = self._get_folder_by_iter(store, folder_iter)
-
-        can_del = bool(folder and (folder not in BASE_FOLDERS))
-
-        action_group = Gtk.ActionGroup.new("menu")
-        actions = [("delete", _("Delete"), Gtk.STOCK_DELETE, can_del),
-                   ("create", _("Create"), Gtk.STOCK_NEW, True),
-                   ("rename", _("Rename"), None, can_del)]
-
-        for action, label, stock, sensitive in actions:
-            new_action = Gtk.Action.new(action, label, None, stock)
-            new_action.set_sensitive(sensitive)
-            new_action.connect("activate", self._menu_handler,
-                               store, folder_iter, view)
-            action_group.add_action(new_action)
-
-        uim = Gtk.UIManager()
-        uim.insert_action_group(action_group, 0)
-        uim.add_ui_from_string(xml)
-        # pylint: disable=no-member
-        uim.get_object("/menu").popup(None, None, None, None,
-                                      event.button, event.time)
-
-    def _mouse_cb(self, view, event):
-        '''
-        Mouse Callback button_press_event Handler.
-
-        :param view: Mouse button widget
-        :type view: :class:`Gtk.TreeView`
-        :param event: Button press event
-        :type event: :class:`Gtk.EventButton`
-        '''
-        if event.button == 1:
-            return self._select_folder(view, event)
-        if event.button == 3:
-            return self._folder_menu(view, event)
-        return None
-
-    # pylint wants a max of 15 local variables
-    # pylint: disable=too-many-locals
-    def _dragged_to(self, view, _ctx, x_coord, y_coord, sel, _info, _ts):
-        '''
-        Dragged to - drag-data-received Handler.
-
-        :param view: Widget getting signal
-        :type view: :class:`Gtk.TreeView`
-        :param _ctx: Context value, unused
-        :type: ctx: :class:`Gtk.DragContext`
-        :param x_coord: Horizontal coordinate of destination
-        :param y_coord: Vertical coordinate of destination
-        :param sel: Selection containing the dragged data
-        :type sel: :class:`Gtk.SelectionData`
-        :param _info: Information registered in the Gtk.TargetList, unused
-        :type _info: int
-        :param _ts: timestamp of when the data was requested, unused
-        :type _ts: int
-        '''
-        (path, _place) = view.get_dest_row_at_pos(x_coord, y_coord)
-
-        text = sel.get_text()
-        byte_data = sel.get_data()
-        text = byte_data.decode('ISO-8859-1').split("\x01")
-        msgs = text[1:]
-
-        src_folder = text[0]
-        dst_folder = self._get_folder_by_iter(view.get_model(),
-                                              view.get_model().get_iter(path))
-
-        if src_folder == dst_folder:
-            return
-
-        dst = MessageFolderInfo(os.path.join(self._folders_path(), dst_folder))
-        src = MessageFolderInfo(os.path.join(self._folders_path(), src_folder))
-
-        for record in msgs:
-            fname, subj, msg_type, read, send, recp = record.split("\0")
-            self.logger.info("_dragged_to Dragged %s from %s into %s",
-                             fname, src_folder, dst_folder)
-            self.logger.info("_dragged_to: %s %s %s %s->%s",
-                             subj, msg_type, read, send, recp)
-
-            try:
-                # Make sure that destination does not exist
-                dst.delete(os.path.basename(fname))
-            except FileNotFoundError:
-                pass
-            newfn = dst.create_msg(os.path.basename(fname))
-            shutil.copy(fname, newfn)
-            src.delete(fname)
-
-            dst.set_msg_read(fname, read == "True")
-            dst.set_msg_subject(fname, subj)
-            dst.set_msg_type(fname, msg_type)
-            dst.set_msg_sender(fname, send)
-            dst.set_msg_recip(fname, recp)
-
-    def _folder_rename(self, _render, path, new_text, store):
-        '''
-        Folder rename edited handler.
-
-        :param _render: Rendering widget
-        :type _render: :class:`Gtk.CellRendererText`
-        :param path: Path identifying edited cell
-        :type path: str
-        :param new_text: New text
-        :type new_text: str
-        :param store: TreeStore for data
-        :type store: :class:`Gtk.TreeStore`
-        '''
-        folder_iter = store.get_iter(path)
-        orig = store.get(folder_iter, 0)[0]
-        if orig == new_text:
-            return
-        if orig in BASE_FOLDERS:
-            return
-        info = self.get_folder(self._get_folder_by_iter(store, folder_iter))
-        try:
-            info.rename(new_text)
-        except OSError as err:
-            display_error("Unable to rename: %s" % err)
-            return
-
-        store.set(iter, 0, new_text)
-
-
-ML_COL_ICON = 0
-ML_COL_SEND = 1
-ML_COL_SUBJ = 2
-ML_COL_TYPE = 3
-ML_COL_DATE = 4
-ML_COL_FILE = 5
-ML_COL_READ = 6
-ML_COL_RECP = 7
-
-
-class MessageList(MainWindowElement):
-    '''
-    Message List.
-
-    :param wtree: Window Tree Object.
-    :param config: Configuration object
-    :type config: :class:`DratsConfig`
-    '''
-
-    __gsignals__ = {"prompt-send-form" : (GObject.SignalFlags.RUN_LAST,
-                                          GObject.TYPE_NONE,
-                                          (GObject.TYPE_STRING,)),
-                    "reply-form" : (GObject.SignalFlags.RUN_LAST,
-                                    GObject.TYPE_NONE,
-                                    (GObject.TYPE_STRING,)),
-                    "delete-form" : (GObject.SignalFlags.RUN_LAST,
-                                     GObject.TYPE_NONE,
-                                     (GObject.TYPE_STRING,)),
-                    }
-
-    # pylint wants a max of 50 statements for methods or functions
-    # pylint: disable=too-many-statements
-    def __init__(self, wtree, config):
-        MainWindowElement.__init__(self, wtree, config, "msg")
-
-        self.logger = logging.getLogger("MessageList")
-        msglist = self._get_widget("msglist")
-
-        self.store = Gtk.ListStore(GObject.TYPE_OBJECT,
-                                   GObject.TYPE_STRING,
-                                   GObject.TYPE_STRING,
-                                   GObject.TYPE_STRING,
-                                   GObject.TYPE_INT,
-                                   GObject.TYPE_STRING,
-                                   GObject.TYPE_BOOLEAN,
-                                   GObject.TYPE_STRING)
-        msglist.set_model(self.store)
-        msglist.get_selection().set_mode(Gtk.SelectionMode.MULTIPLE)
-        msglist.enable_model_drag_source(Gdk.ModifierType.BUTTON1_MASK,
-                                         [("text/d-rats_message", 0, 0)],
-                                         Gdk.DragAction.DEFAULT|
-                                         Gdk.DragAction.MOVE)
-        msglist.connect("drag-data-get", self._dragged_from)
-
-        col = Gtk.TreeViewColumn("", Gtk.CellRendererPixbuf(), pixbuf=0)
-        msglist.append_column(col)
-
-        def bold_if_unread(_col, rend, model, msg_iter, cnum):
-            val, read, = model.get(msg_iter, cnum, ML_COL_READ)
-            if not val:
-                val = ""
-            if not read:
-                val = val.replace("&", "&amp;")
-                val = val.replace("<", "&lt;")
-                val = val.replace(">", "&gt;")
-                rend.set_property("markup", "<b>%s</b>" % val)
-
-        renderer = Gtk.CellRendererText()
-        renderer.set_property("ellipsize", Pango.EllipsizeMode.END)
-        col = Gtk.TreeViewColumn(_("Sender"), renderer, text=ML_COL_SEND)
-        col.set_cell_data_func(renderer, bold_if_unread, ML_COL_SEND)
-        col.set_sort_column_id(ML_COL_SEND)
-        col.set_sizing(Gtk.TreeViewColumnSizing.AUTOSIZE)
-        col.set_resizable(True)
-        msglist.append_column(col)
-
-        renderer = Gtk.CellRendererText()
-        renderer.set_property("ellipsize", Pango.EllipsizeMode.END)
-        col = Gtk.TreeViewColumn(_("Recipient"), renderer, text=ML_COL_RECP)
-        col.set_cell_data_func(renderer, bold_if_unread, ML_COL_RECP)
-        col.set_sort_column_id(ML_COL_RECP)
-        col.set_sizing(Gtk.TreeViewColumnSizing.AUTOSIZE)
-        col.set_resizable(True)
-        msglist.append_column(col)
-
-        renderer = Gtk.CellRendererText()
-        renderer.set_property("ellipsize", Pango.EllipsizeMode.END)
-        col = Gtk.TreeViewColumn(_("Subject"), renderer, text=ML_COL_SUBJ)
-        col.set_cell_data_func(renderer, bold_if_unread, ML_COL_SUBJ)
-        col.set_expand(True)
-        col.set_sort_column_id(ML_COL_SUBJ)
-        col.set_sizing(Gtk.TreeViewColumnSizing.AUTOSIZE)
-        col.set_resizable(True)
-        msglist.append_column(col)
-
-        renderer = Gtk.CellRendererText()
-        renderer.set_property("ellipsize", Pango.EllipsizeMode.END)
-        col = Gtk.TreeViewColumn(_("Type"), renderer, text=ML_COL_TYPE)
-        col.set_cell_data_func(renderer, bold_if_unread, ML_COL_TYPE)
-        col.set_sort_column_id(ML_COL_TYPE)
-        col.set_resizable(True)
-        msglist.append_column(col)
-
-        def render_date(_col, rend, model, msg_iter, _data):
-            time_stamp, read = model.get(msg_iter, ML_COL_DATE, ML_COL_READ)
-            stamp = datetime.fromtimestamp(
-                time_stamp).strftime("%H:%M:%S %Y-%m-%d")
-            if read:
-                rend.set_property("text", stamp)
-            else:
-                rend.set_property("markup", "<b>%s</b>" % stamp)
-
-        renderer = Gtk.CellRendererText()
-        renderer.set_property("ellipsize", Pango.EllipsizeMode.END)
-        col = Gtk.TreeViewColumn(_("Date"), renderer, text=ML_COL_DATE)
-        col.set_cell_data_func(renderer, render_date)
-        col.set_sort_column_id(ML_COL_DATE)
-        col.set_resizable(True)
-        msglist.append_column(col)
-
-        msglist.connect("row-activated", self._open_msg)
-        self.store.set_sort_column_id(ML_COL_DATE, Gtk.SortType.DESCENDING)
-
-        self.message_pixbuf = self._config.ship_img("message.png")
-        self.unread_pixbuf = self._config.ship_img("msg-markunread.png")
-        self.current_info = None
-
-    def _folder_path(self, folder):
-        path = os.path.join(self._config.platform.config_dir(),
-                            "messages",
-                            folder)
-        if not os.path.isdir(path):
-            return None
-        return path
-
-    def open_msg(self, filename, editable, call_back=None, cbdata=None):
-        '''
-        Open a message.
-
-        :param filename: Filename for message
-        :type filename: str
-        :param editable: If message can be edited
-        :type editable: bool
-        :param call_back: Callback for message
-        :type call_back: function(:class:`Gtk.ResponseType`, any)
-        :param cbdata: Call back data
-        :type cbdata: any
-        :returns: response
-        :rtype: :class:`Gtk.ResponseType`
-        '''
-        if not msgrouting.msg_lock(filename):
-            display_error(_("Unable to open: message in use by another task"))
-            return Gtk.ResponseType.CANCEL
-
-        parent = self._wtree.get_object("mainwindow")
-        form = formgui.FormDialog(_("Form"), filename,
-                                  config=self._config, parent=parent)
-        form.configure(self._config)
-
-        def form_done(dlg, response, msg_info):
-            '''
-            Form Done response handler.
-
-            :param dlg: Message Info dialog
-            :param dlg: :class:`formgui.FormDialog`
-            :param response: response from dialog
-            :type response: :class:`Gtk.ResponseType`
-            :param msg_info: Message info
-            :type msg_info: :class:`MessageInfo`
-            '''
-            saveable_actions = [formgui.RESPONSE_SAVE,
-                                formgui.RESPONSE_SEND,
-                                formgui.RESPONSE_SEND_VIA,
-                                ]
-            dlg.hide()
-            dlg.update_dst()
-            filename = msg_info.filename
-            info = msg_info.info
-            if msgrouting.msg_is_locked(filename):
-                msgrouting.msg_unlock(filename)
-            if response in saveable_actions:
-                self.logger.info("open_msg: Saving to %s", filename)
-                dlg.save_to(filename)
-            else:
-                self.logger.info("open_msg : Not saving")
-            dlg.destroy()
-            self.refresh(filename)
-            if call_back:
-                call_back(response, cbdata)
-            if response == formgui.RESPONSE_SEND:
-                self.move_message(info, filename, _("Outbox"))
-            elif response == formgui.RESPONSE_SEND_VIA:
-                filename = self.move_message(info, filename, _("Outbox"))
-                self.emit("prompt-send-form", filename)
-            elif response == formgui.RESPONSE_REPLY:
-                self.emit("reply-form", filename)
-            elif response == formgui.RESPONSE_DELETE:
-                self.emit("delete-form", filename)
-
-        form.build_gui(editable)
-        form.show()
-        msg_info = MessageInfo(filename, self.current_info)
-        form.connect("response", form_done, msg_info)
-        return Gtk.ResponseType.OK
-
-    def _open_msg(self, view, path, _col):
-        store = view.get_model()
-        msg_iter = store.get_iter(path)
-        path, = store.get(msg_iter, ML_COL_FILE)
-
-        def close_msg_cb(response, info):
-            if self.current_info == info:
-                msg_iter = self.iter_from_fn(path)
-                self.logger.info("_open_msg: Updating iter for close %s",
-                                 msg_iter)
-                if msg_iter:
-                    self._update_message_info(msg_iter)
-            else:
-                self.logger.info("_open_msg: Not current, not updating")
-
-        editable = "Outbox" in path or "Drafts" in path # Dirty hack
-        self.open_msg(path, editable, close_msg_cb, self.current_info)
-        self.current_info.set_msg_read(path, True)
-        msg_iter = self.iter_from_fn(path)
-        self.logger.info("_open_msg: Updating iter %s", msg_iter)
-        if msg_iter:
-            self._update_message_info(msg_iter)
-
-    def _dragged_from(self, view, _ctx, sel, _info, _ts):
-        '''
-        Dragged from.
-
-        :param view: Widget getting signal
-        :type view: :class:`Gtk.Widget`
-        :param _ctx: Context value, unused
-        :type _ctx: :class:`Gtk.DragContext`
-        :param sel: Selection containing the dragged data
-        :type sel: :class:`Gtk.SelectionData`
-        :param _info: Information registered in the Gtk.TargetList, unused
-        :param _ts: timestamp of when the data was requested, unused
-        '''
-        store, paths = view.get_selection().get_selected_rows()
-
-        msgs = [os.path.dirname(store[paths[0]][ML_COL_FILE])]
-        for path in paths:
-            data = "%s\0%s\0%s\0%s\0%s\0%s" % (store[path][ML_COL_FILE],
-                                               store[path][ML_COL_SUBJ],
-                                               store[path][ML_COL_TYPE],
-                                               store[path][ML_COL_READ],
-                                               store[path][ML_COL_SEND],
-                                               store[path][ML_COL_RECP])
-            msgs.append(data)
-
-        data = "\x01".join(msgs)
-        byte_data = data.encode('ISO-8859-1')
-        sel.set(sel.get_target(), 0, byte_data)
-        GLib.idle_add(self.refresh)
-
-
-    def _update_message_info(self, msg_iter, force=False):
-        fname, = self.store.get(msg_iter, ML_COL_FILE)
-
-        subj = self.current_info.get_msg_subject(fname)
-        read = self.current_info.get_msg_read(fname)
-        if subj == _("Unknown") or force:
-            # Not registered, so update the registry
-            form = formgui.FormFile(fname)
-            self.current_info.set_msg_type(fname, form.ident)
-            self.current_info.set_msg_read(fname, read)
-            self.current_info.set_msg_subject(fname, form.get_subject_string())
-            self.current_info.set_msg_sender(fname, form.get_sender_string())
-            self.current_info.set_msg_recip(fname, form.get_recipient_string())
-
-        time_stamp = os.stat(fname).st_ctime
-        if read:
-            icon = self.message_pixbuf
-        else:
-            icon = self.unread_pixbuf
-        self.store.set(msg_iter,
-                       ML_COL_ICON, icon,
-                       ML_COL_SEND, self.current_info.get_msg_sender(fname),
-                       ML_COL_RECP, self.current_info.get_msg_recip(fname),
-                       ML_COL_SUBJ, self.current_info.get_msg_subject(fname),
-                       ML_COL_TYPE, self.current_info.get_msg_type(fname),
-                       ML_COL_DATE, time_stamp,
-                       ML_COL_READ, read)
-
-    def iter_from_fn(self, file_name):
-        '''
-        Iterate from file name.
-
-        :param file_name: File Name to lookup
-        :type file_name: str
-        :returns: Iterated file name with each call
-        :rtype: :class:`Gtk.TreeIter`
-        '''
-        fn_iter = self.store.get_iter_first()
-        while fn_iter:
-            _fn, = self.store.get(fn_iter, ML_COL_FILE)
-            if _fn == file_name:
-                break
-            fn_iter = self.store.iter_next(fn_iter)
-
-        return fn_iter
-
-    def refresh(self, file_name=None):
-        '''
-        Refresh the current folder or optional filename.
-
-        :param file_name: File name, default None
-        :type file_name: str
-        '''
-        if file_name is None:
-            self.store.clear()
-            for msg in self.current_info.files():
-                msg_iter = self.store.append()
-                self.store.set(msg_iter, ML_COL_FILE, msg)
-                self._update_message_info(msg_iter)
-        else:
-            msg_iter = self.iter_from_fn(file_name)
-            if not msg_iter:
-                msg_iter = self.store.append()
-                self.store.set(msg_iter,
-                               ML_COL_FILE, file_name)
-
-            self._update_message_info(msg_iter, True)
-
-    def open_folder(self, path):
-        '''
-        Open a folder by path.
-
-        :param path: Folder to open
-        :type path: str
-        '''
-        self.current_info = MessageFolderInfo(self._folder_path(path))
-        self.refresh()
-
-    def delete_selected_messages(self):
-        '''Delete Selected Messages.'''
-        msglist = self._get_widget("msglist")
-
-        iters = []
-        (store, paths) = msglist.get_selection().get_selected_rows()
-        for path in paths:
-            iters.append(store.get_iter(path))
-
-        for msg_iter in iters:
-            file_name, = store.get(msg_iter, ML_COL_FILE)
-            store.remove(msg_iter)
-            self.current_info.delete(file_name)
-
-    def move_message(self, info, path, new_folder):
-        '''
-        Move message into folder.
-
-        :param info: Message information
-        :type info: :class:`MessageFolderInfo`
-        :param path: Source folder
-        :type path: str
-        :param new_folder: Destination Folder
-        :type new_folder: str
-        :returns: The new folder name on success, the source folder on failure
-        :rtype: str
-        '''
-        dest = MessageFolderInfo(self._folder_path(new_folder))
-        try:
-            newfn = dest.create_msg(os.path.basename(path))
-        except DuplicateSectionError:
-            # Same folder, or duplicate message id
-            return path
-
-        self.logger.info("move_message Moving %s -> %s", path, newfn)
-        shutil.copy(path, newfn)
-        info.delete(path)
-
-        if info == self.current_info:
-            self.refresh()
-
-        return newfn
-
-    def move_selected_messages(self, folder):
-        '''
-        Move selected messages into folder.
-
-        :param folder: Destination folder
-        :type folder: str
-        '''
-        for msg in self.get_selected_messages():
-            self.move_message(self.current_info, msg, folder)
-
-    def get_selected_messages(self):
-        '''
-        Get selected messages.
-
-        :returns: List of selected messages
-        :rtype: list
-        '''
-        msglist = self._get_widget("msglist")
-
-        selected = []
-        (store, paths) = msglist.get_selection().get_selected_rows()
-        for path in paths:
-            selected.append(store[path][ML_COL_FILE])
-
-        return selected
 
 
 class MessagesTab(MainWindowTab):
@@ -1271,16 +155,14 @@ class MessagesTab(MainWindowTab):
         self._messages.open_msg(newfn, True,
                                 close_msg_cb, self._messages.current_info)
 
-    # pylint wants a max of 15 local variables
-    # pylint: disable=too-many-locals
-    def _rpl_msg(self, _button, file_name=None):
+    def _set_rpl_msg_fields(self, old_form, new_form):
         '''
-        Reply to Message reply-form and clicked handler.
+        Set Reply Message Fields.
 
-        :param _button: Widget that was signaled, unused
-        :type _button: :class:`MessageList`, :class:`Gtk.MenuToolButton`
-        :param file_name: Filename for message, default None
-        :type file_name: str
+        :param old_form: Incoming form for message
+        :type old_form: :class:`formgui.FormFile`
+        :param new_form: Reply message form
+        :type new_form: :class:`formgui.FormFile`
         '''
         def subj_reply(subj):
             if "RE:" in subj.upper():
@@ -1300,6 +182,25 @@ class MessagesTab(MainWindowTab):
             ("_auto_sender", "_auto_recip", None),
             ]
 
+        for s_field, d_field, x_field in save_fields:
+            old_val = old_form.get_field_value(s_field)
+            if not old_val:
+                continue
+
+            if x_field:
+                new_form.set_field_value(d_field, x_field(old_val))
+            else:
+                new_form.set_field_value(d_field, old_val)
+
+    def _rpl_msg(self, _button, file_name=None):
+        '''
+        Reply to Message reply-form and clicked handler.
+
+        :param _button: Widget that was signaled, unused
+        :type _button: :class:`MessageList`, :class:`Gtk.MenuToolButton`
+        :param file_name: Filename for message, default None
+        :type file_name: str
+        '''
         if not file_name:
             try:
                 sel = self._messages.get_selected_messages()
@@ -1315,44 +216,36 @@ class MessagesTab(MainWindowTab):
         current = self._messages.current_info.name()
         self._folders.select_folder(_("Drafts"))
 
-        oform = formgui.FormFile(file_name)
+        old_form = formgui.FormFile(file_name)
         tmpl = os.path.join(self._config.form_source_dir(),
-                            "%s.xml" % oform.ident)
+                            "%s.xml" % old_form.ident)
 
-        nform = formgui.FormFile(tmpl)
-        nform.add_path_element(self._config.get("user", "callsign"))
+        new_form = formgui.FormFile(tmpl)
+        new_form.add_path_element(self._config.get("user", "callsign"))
 
         try:
-            for s_field, d_field, x_field in save_fields:
-                old_val = oform.get_field_value(s_field)
-                if not old_val:
-                    continue
-
-                if x_field:
-                    nform.set_field_value(d_field, x_field(old_val))
-                else:
-                    nform.set_field_value(d_field, old_val)
+            self._set_rpl_msg_fields(old_form, new_form)
 
         except formgui.FormguiFileMultipleIds:
             self.logger.info("_rpl_msg: Failed to do reply",
                              exc_info=True, stack_info=True)
             return
 
-        if ";" in oform.get_path_dst():
-            rpath = ";".join(reversed(oform.get_path()[:-1]))
+        if ";" in old_form.get_path_dst():
+            rpath = ";".join(reversed(old_form.get_path()[:-1]))
             self.logger.info("_rpl_msg: rpath: %s (%s)",
-                             rpath, oform.get_path())
-            nform.set_path_dst(rpath)
+                             rpath, old_form.get_path())
+            new_form.set_path_dst(rpath)
         else:
-            nform.set_path_dst(oform.get_path_src())
+            new_form.set_path_dst(old_form.get_path_src())
 
         call = self._config.get("user", "callsign")
-        nform.set_path_src(call)
-        nform.set_path_mid(mkmsgid(call))
+        new_form.set_path_src(call)
+        new_form.set_path_mid(mkmsgid(call))
 
         tstamp = time.strftime("form_%m%d%Y_%H%M%S.xml")
         newfn = self._messages.current_info.create_msg(tstamp)
-        nform.save_to(newfn)
+        new_form.save_to(newfn)
 
         def close_msg_cb(response, info):
             if self._messages.current_info == info:
